Initialization and validation really ought to all live together. Fixed!
